requirements.txt
numpy
scipy
opencv-python-headless
matplotlib
joblib
tqdm
pyvista
pyyaml
